+++ docs/CHANGELOG.md
@@ -17,6 +17,7 @@ It is still possible to explicitly choose a specific build backend with `--build
 
 * Abort migration early if uv executable is required but not found ([#558](https://github.com/mkniewallner/migrate-to-uv/pull/558))
 * Improve dependencies locking wording ([#609](https://github.com/mkniewallner/migrate-to-uv/pull/609))
+* Add `--keep-current-build-backend` flag ([#614](https://github.com/mkniewallner/migrate-to-uv/pull/614))
 * [poetry] Choose build backend based on distribution complexity ([#597](https://github.com/mkniewallner/migrate-to-uv/pull/597))
 * [poetry] Automatically add Python classifiers for packages based on `python` specifier ([#606](https://github.com/mkniewallner/migrate-to-uv/pull/606))